+++ aws-java-sdk-emrcontainers/src/main/java/com/amazonaws/services/emrcontainers/model/transform/StartJobRunRequestMarshaller.java
@@ -48,6 +48,10 @@ public class StartJobRunRequestMarshaller {
             .marshallLocation(MarshallLocation.PAYLOAD).marshallLocationName("configurationOverrides").build();
     private static final MarshallingInfo<Map> TAGS_BINDING = MarshallingInfo.builder(MarshallingType.MAP).marshallLocation(MarshallLocation.PAYLOAD)
             .marshallLocationName("tags").build();
+    private static final MarshallingInfo<String> JOBTEMPLATEID_BINDING = MarshallingInfo.builder(MarshallingType.STRING)
+            .marshallLocation(MarshallLocation.PAYLOAD).marshallLocationName("jobTemplateId").build();
+    private static final MarshallingInfo<Map> JOBTEMPLATEPARAMETERS_BINDING = MarshallingInfo.builder(MarshallingType.MAP)
+            .marshallLocation(MarshallLocation.PAYLOAD).marshallLocationName("jobTemplateParameters").build();
 
     private static final StartJobRunRequestMarshaller instance = new StartJobRunRequestMarshaller();
 
@@ -73,6 +77,8 @@ public class StartJobRunRequestMarshaller {
             protocolMarshaller.marshall(startJobRunRequest.getJobDriver(), JOBDRIVER_BINDING);
             protocolMarshaller.marshall(startJobRunRequest.getConfigurationOverrides(), CONFIGURATIONOVERRIDES_BINDING);
             protocolMarshaller.marshall(startJobRunRequest.getTags(), TAGS_BINDING);
+            protocolMarshaller.marshall(startJobRunRequest.getJobTemplateId(), JOBTEMPLATEID_BINDING);
+            protocolMarshaller.marshall(startJobRunRequest.getJobTemplateParameters(), JOBTEMPLATEPARAMETERS_BINDING);
         } catch (Exception e) {
             throw new SdkClientException("Unable to marshall request to JSON: " + e.getMessage(), e);
         }