+++ aws-java-sdk-logs/src/main/java/com/amazonaws/services/logs/AWSLogsClient.java
@@ -502,6 +502,13 @@ public class AWSLogsClient extends AmazonWebServiceClient implements AWSLogs {
      * the S3 bucket that you specify as the destination.
      * </p>
      * <p>
+     * Exporting log data to Amazon S3 buckets that are encrypted by KMS is supported. Exporting log data to Amazon S3
+     * buckets that have S3 Object Lock enabled with a retention period is also supported.
+     * </p>
+     * <p>
+     * Exporting to S3 buckets that are encrypted with AES-256 is supported.
+     * </p>
+     * <p>
      * This is an asynchronous call. If all the required information is provided, this operation initiates an export
      * task and responds with the ID of the task. After the task has started, you can use <a
      * href="https://docs.aws.amazon.com/AmazonCloudWatchLogs/latest/APIReference/API_DescribeExportTasks.html"
@@ -515,10 +522,12 @@ public class AWSLogsClient extends AmazonWebServiceClient implements AWSLogs {
      * data for each export task, you can specify a prefix to be used as the Amazon S3 key prefix for all exported
      * objects.
      * </p>
+     * <note>
      * <p>
-     * Exporting to S3 buckets that are encrypted with AES-256 is supported. Exporting to S3 buckets encrypted with
-     * SSE-KMS is not supported.
+     * Time-based sorting on chunks of log data inside an exported file is not guaranteed. You can sort the exported log
+     * fild data by using Linux utilities.
      * </p>
+     * </note>
      * 
      * @param createExportTaskRequest
      * @return Result of the CreateExportTask operation returned by the service.
@@ -1951,6 +1960,9 @@ public class AWSLogsClient extends AmazonWebServiceClient implements AWSLogs {
      * filter pattern, a time range, and the name of the log stream.
      * </p>
      * <p>
+     * You must have the <code>logs;FilterLogEvents</code> permission to perform this operation.
+     * </p>
+     * <p>
      * By default, this operation returns as many log events as can fit in 1 MB (up to 10,000 log events) or all the
      * events found within the time range that you specify. If the results include a token, then there are more log
      * events available, and you can get additional results by specifying the token in a subsequent call. This operation
@@ -2938,6 +2950,19 @@ public class AWSLogsClient extends AmazonWebServiceClient implements AWSLogs {
      * Sets the retention of the specified log group. A retention policy allows you to configure the number of days for
      * which to retain log events in the specified log group.
      * </p>
+     * <note>
+     * <p>
+     * CloudWatch Logs doesn’t immediately delete log events when they reach their retention setting. It typically takes
+     * up to 72 hours after that before log events are deleted, but in rare situations might take longer.
+     * </p>
+     * <p>
+     * This means that if you change a log group to have a longer retention setting when it contains log events that are
+     * past the expiration date, but haven’t been actually deleted, those log events will take up to 72 hours to be
+     * deleted after the new retention date is reached. To make sure that log data is deleted permanently, keep a log
+     * group at its lower retention setting until 72 hours has passed after the end of the previous retention period, or
+     * you have confirmed that the older log events are deleted.
+     * </p>
+     * </note>
      * 
      * @param putRetentionPolicyRequest
      * @return Result of the PutRetentionPolicy operation returned by the service.
@@ -3114,6 +3139,10 @@ public class AWSLogsClient extends AmazonWebServiceClient implements AWSLogs {
      * <p>
      * Queries time out after 15 minutes of execution. If your queries are timing out, reduce the time range being
      * searched or partition your query into a number of queries.
+     * </p>
+     * <p>
+     * You are limited to 20 concurrent CloudWatch Logs insights queries, including queries that have been added to
+     * dashboards.
      * </p>
      * 
      * @param startQueryRequest